fix int/size_t type errors

+++ wrap/utils/DefReader.d
@@ -115,7 +115,7 @@ public class DefReader
 		
 		if ( line.length > 0 )
 		{
-			int pos = std.string.indexOf(line, ':');
+			size_t pos = std.string.indexOf(line, ':');
 			if ( pos > 0 )
 			{
 				key = std.string.strip(line[0 .. pos]);
